added some changes to the repository
