+++ pyext/src/Files.cmake
@@ -1,0 +1,3 @@
+set(pyfiles "restraints/strings.py")
+set(cppfiles "")
+set(cudafiles "")
